Implement a better atom matching algorithm
Compare heavy atoms by their closeness in space.
Then compare hydrogen atoms attached to the heavy atoms based
on whether the heavy atom neighbors are all matched or not.
If all heavy atom neighbors match then the hydrogens are matched to each
other.
If some of the heavy atom neighbors don't have matches then the
hydrogens are not matched to each other.