Merged main into llvm14 branch
This was a painful merge - there were lots of changes.
It builds.  I hope everything works (gulp).

This merge involves lots of changes to kinematics code to simplify it.
I put too much kinematics code in C++.
I'm simplifying it now so that there is only
a small core in C++ and the rest will be in Common Lisp.