fix some build problems
Specifically: in elements.h the translate:: is redundant since we
are already in the translate namespace, and that causes a warning.

In the other file we -> into a Symbol_sp even though symbol.h was
not directly included. It was included transiently but I rearranged
that away in Clasp in a branch.